Revert commits which move `justification` into primitives
This reverts commit 03a381f0bc4a8dbe4785c30d42ab252a06ba876c.
