+++ modules/substrate/Cargo.toml
@@ -9,7 +9,6 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 
 [dependencies]
 bp-header-chain = { path = "../../primitives/header-chain", default-features = false }
-bp-substrate = { path = "../../primitives/substrate", default-features = false }
 finality-grandpa = { version = "0.12.3", default-features = false }
 hash-db = { version = "0.15.2", default-features = false }
 serde = { version = "1.0", optional = true }
@@ -60,9 +59,6 @@ default-features = false
 git = "https://github.com/paritytech/substrate/"
 
 # Dev Dependencies
-# [dev-dependencies]
-# bp-substrate = { path = "../../primitives/substrate", default-features = false }
-
 [dev-dependencies.sp-io]
 version = "2.0.0"
 tag = 'v2.0.0'
