+++ modules/substrate/src/justification.rs
@@ -121,7 +121,7 @@ where
 /// (so not our chain) have been finalized correctly.
 #[derive(Decode, RuntimeDebug)]
 #[cfg_attr(test, derive(codec::Encode))]
-pub struct GrandpaJustification<Header: HeaderT> {
+pub(crate) struct GrandpaJustification<Header: HeaderT> {
 	round: u64,
 	commit: finality_grandpa::Commit<Header::Hash, Header::Number, AuthoritySignature, AuthorityId>,
 	votes_ancestries: Vec<Header>,
@@ -174,18 +174,81 @@ where
 }
 
 #[cfg(test)]
-pub mod tests {
+pub(crate) mod tests {
 	use super::*;
-	use crate::test_helpers::*;
+	use crate::mock::helpers::*;
 	use codec::Encode;
+	use sp_core::H256;
 	use sp_finality_grandpa::{AuthorityId, AuthorityWeight};
 	use sp_keyring::Ed25519Keyring;
-	use sp_runtime::testing::{Header, H256};
 
 	const TEST_GRANDPA_ROUND: u64 = 1;
 	const TEST_GRANDPA_SET_ID: SetId = 1;
 
-	fn make_justification_for_header_1() -> GrandpaJustification<Header> {
+	pub(crate) fn signed_precommit(
+		signer: Ed25519Keyring,
+		target: HeaderId,
+		round: u64,
+		set_id: SetId,
+	) -> finality_grandpa::SignedPrecommit<H256, u64, AuthoritySignature, AuthorityId> {
+		let precommit = finality_grandpa::Precommit {
+			target_hash: target.0,
+			target_number: target.1,
+		};
+		let encoded = sp_finality_grandpa::localized_payload(
+			round,
+			set_id,
+			&finality_grandpa::Message::Precommit(precommit.clone()),
+		);
+		let signature = signer.sign(&encoded[..]).into();
+		finality_grandpa::SignedPrecommit {
+			precommit,
+			signature,
+			id: signer.public().into(),
+		}
+	}
+
+	pub(crate) fn make_justification_for_header(
+		header: &TestHeader,
+		round: u64,
+		set_id: SetId,
+		authorities: &[(AuthorityId, AuthorityWeight)],
+	) -> GrandpaJustification<TestHeader> {
+		let (target_hash, target_number) = (header.hash(), *header.number());
+		let mut precommits = vec![];
+		let mut votes_ancestries = vec![];
+
+		// We want to make sure that the header included in the vote ancestries
+		// is actually related to our target header
+		let mut precommit_header = test_header(target_number + 1);
+		precommit_header.parent_hash = target_hash;
+
+		// I'm using the same header for all the voters since it doesn't matter as long
+		// as they all vote on blocks _ahead_ of the one we're interested in finalizing
+		for (id, _weight) in authorities.iter() {
+			let signer = extract_keyring(&id);
+			let precommit = signed_precommit(
+				signer,
+				(precommit_header.hash(), *precommit_header.number()),
+				round,
+				set_id,
+			);
+			precommits.push(precommit);
+			votes_ancestries.push(precommit_header.clone());
+		}
+
+		GrandpaJustification {
+			round,
+			commit: finality_grandpa::Commit {
+				target_hash,
+				target_number,
+				precommits,
+			},
+			votes_ancestries,
+		}
+	}
+
+	pub(crate) fn make_justification_for_header_1() -> GrandpaJustification<TestHeader> {
 		make_justification_for_header(
 			&test_header(1),
 			TEST_GRANDPA_ROUND,
@@ -197,7 +260,7 @@ pub mod tests {
 	#[test]
 	fn justification_with_invalid_encoding_rejected() {
 		assert_eq!(
-			verify_justification::<Header>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &[],),
+			verify_justification::<TestHeader>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &[],),
 			Err(Error::JustificationDecode),
 		);
 	}
@@ -205,7 +268,7 @@ pub mod tests {
 	#[test]
 	fn justification_with_invalid_target_rejected() {
 		assert_eq!(
-			verify_justification::<Header>(
+			verify_justification::<TestHeader>(
 				header_id(2),
 				TEST_GRANDPA_SET_ID,
 				voter_set(),
@@ -221,7 +284,7 @@ pub mod tests {
 		justification.commit.precommits.clear();
 
 		assert_eq!(
-			verify_justification::<Header>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &justification.encode(),),
+			verify_justification::<TestHeader>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &justification.encode(),),
 			Err(Error::InvalidJustificationCommit),
 		);
 	}
@@ -232,7 +295,7 @@ pub mod tests {
 		justification.commit.precommits[0].signature = Default::default();
 
 		assert_eq!(
-			verify_justification::<Header>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &justification.encode(),),
+			verify_justification::<TestHeader>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &justification.encode(),),
 			Err(Error::InvalidAuthoritySignature),
 		);
 	}
@@ -243,7 +306,7 @@ pub mod tests {
 		justification.votes_ancestries.push(test_header(10));
 
 		assert_eq!(
-			verify_justification::<Header>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &justification.encode(),),
+			verify_justification::<TestHeader>(header_id(1), TEST_GRANDPA_SET_ID, voter_set(), &justification.encode(),),
 			Err(Error::InvalidPrecommitAncestries),
 		);
 	}
@@ -251,7 +314,7 @@ pub mod tests {
 	#[test]
 	fn valid_justification_accepted() {
 		assert_eq!(
-			verify_justification::<Header>(
+			verify_justification::<TestHeader>(
 				header_id(1),
 				TEST_GRANDPA_SET_ID,
 				voter_set(),
