+++ modules/substrate/src/lib.rs
@@ -45,6 +45,7 @@ use sp_runtime::traits::{
 };
 use sp_std::{fmt::Debug, marker::PhantomData, prelude::*, str::FromStr};
 
+mod justification;
 mod storage;
 mod storage_proof;
 mod verifier;
