+++ primitives/substrate/Cargo.toml
@@ -8,27 +8,13 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-finality-grandpa = { version = "0.12.3", default-features = false }
-
-[dependencies.codec]
-package = "parity-scale-codec"
-version = "1.3.1"
-default-features = false
-features = ["derive"]
-
-# Substrate Based Dependencies
-
-[dependencies.frame-support]
-version = "2.0.0"
-tag = 'v2.0.0'
-default-features = false
-git = "https://github.com/paritytech/substrate/"
+serde = { version = "1.0", optional = true }
 
 [dependencies.sp-finality-grandpa]
 version = "2.0.0"
 tag = 'v2.0.0'
 default-features = false
-git = "https://github.com/paritytech/substrate/"
+git = "https://github.com/paritytech/substrate.git"
 
 [dependencies.sp-runtime]
 version = "2.0.0"
@@ -36,27 +22,15 @@ tag = 'v2.0.0'
 default-features = false
 git = "https://github.com/paritytech/substrate/"
 
-[dependencies.sp-std]
-version = "2.0.0"
-tag = 'v2.0.0'
+[dependencies.parity-scale-codec]
+version = "1.3.1"
 default-features = false
-git = "https://github.com/paritytech/substrate/"
-
-# Dev Dependencies
-
-[dev-dependencies.sp-keyring]
-version = "2.0.0"
-tag = 'v2.0.0'
-default-features = false
-git = "https://github.com/paritytech/substrate/"
+features = ["derive"]
 
 [features]
 default = ["std"]
 std = [
-	"codec/std",
-	"finality-grandpa/std",
-	"frame-support/std",
-	"sp-finality-grandpa/std",
+	"serde/std",
 	"sp-runtime/std",
-	"sp-std/std",
+	"sp-finality-grandpa/std",
 ]
